refactor: move ExecutionMode out of the events module into common types
