+++ crates/common/common_utils/src/types.rs
@@ -471,3 +471,22 @@ impl FromStr for SemanticVersion {
         )?))
     }
 }
+
+/// Primary execution or shadow mirror, derived from the `x-shadow-mode` metadata flag.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum ExecutionMode {
+    Primary,
+    Shadow,
+}
+
+impl ExecutionMode {
+    /// Map the boolean shadow flag (from the `x-shadow-mode` metadata) to an execution mode.
+    pub fn from_shadow_flag(shadow_mode: bool) -> Self {
+        if shadow_mode {
+            Self::Shadow
+        } else {
+            Self::Primary
+        }
+    }
+}
