+++ crates/common/external-services/src/service.rs
@@ -202,7 +202,7 @@ use common_utils::events::{Event, EventConfig, FlowName};
 // Only the injector-client build compiles `create_event` (the outbound connector
 // call path), which is the sole user of these event-dimension enums.
 #[cfg(feature = "injector-client")]
-use common_utils::events::ExecutionMode;
+use common_utils::types::ExecutionMode;
 #[cfg(feature = "injector-client")]
 // TokenData is now imported from hyperswitch_injector
 use common_utils::{consts, emit_event_with_config};
